moved elements to correct folder instead of nested deep in unlogical places
