--- src/Kartlag/AktiveKartlag/EkspandertMeny/Visualisering/Skalanivå/Skalanivå.js
@@ -1,23 +1,0 @@
-import React, { Component } from "react";
-
-class Skalanivå extends Component {
-  render() {
-    
-    /*
-    Legg til: 
-
-    * Mulighet for å velge skala på det man ser på for høytliggende elementer,
-      vil påvirke underelementene. 
-
-    */
-
-    return (
-      <>
-        
-        
-      </>
-    );
-  }
-}
-
-export default Skalanivå;
