+++ src/Kartlag/CurrentElement/CurrentElement.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import CurrentLayerSettings from "../AktiveKartlag/EkspandertMeny/CurrentLayerSettings";
-import Navigeringsliste from "./Navigeringsliste";
+import CurrentLayerSettings from "./CurrentLayerSettings";
+import LayerSubElementListe from "./LayerSubElementListe";
 import FavoriteButton from "../Buttons/FavoriteButton";
 import SettingsButton from "../Buttons/SettingsButton";
 import InfoButton from "../Buttons/InfoButton";
@@ -66,7 +66,7 @@ Sidebarmeny-navigeringen.
           </>
         )}
 
-        <Navigeringsliste
+        <LayerSubElementListe
           parentkode={meta ? meta.kode : "kode"}
           metadata={meta && meta.barn}
           onNavigate={onNavigate}
