+++ src/Kartlag/CurrentElement/Settings/GradientEllerTypeinndelt.js
