+++ src/Kartlag/Favorites/FavoriteLayerElement.js
